+++ client/src/components/logbooks/LogDetail/style/ItemRowCard.style.ts
