+++ client/src/components/logbooks/LogDetail/style/ItemRows.style.ts
