+++ client/src/components/logbooks/LogDetail/style/LogDetail.style.ts
