+++ client/src/components/logbooks/LogDetail/style/_common.style.ts
