+++ client/src/components/logbooks/Logbooks/style/MiniLog.style.ts
@@ -1,4 +1,4 @@
-import _commonStyle from "@/components/logbooks/Logbooks/style/_common.style";
+import _commonStyle from "@/components/logbooks/LogDetail/style/_common.style";
 import { colors } from "@/lib/theme/colors";
 import CardStyle from "@/lib/theme/components/Card.style";
 import { flex } from "@/lib/theme/snippets/flex";
